app: Split open dialog into it's own subclass
We shouldn't stuff everything inside the same widget like in this case
the Window.

Closes: https://github.com/p2panda/aardvark/issues/63
