+++ aardvark-app/src/aardvark.gresource.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <gresources>
   <gresource prefix="/org/p2panda/aardvark">
+    <file preprocess="xml-stripblanks">open_dialog/open_dialog.ui</file>
     <file preprocess="xml-stripblanks">window.ui</file>
     <file preprocess="xml-stripblanks">components/zoom_level_selector.ui</file>
     <file preprocess="xml-stripblanks">gtk/help-overlay.ui</file>
