+++ aardvark-app/src/style.css
@@ -8,8 +8,11 @@
 .invite-code-entry {
   border-radius: 15px;
   padding: 15px 20px;
-  min-width: 350px;
   min-height: 120px;
+}
+
+.open-document-button {
+  min-width: 140px;
 }
 
 .user-counter {
